% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_testing.R
\name{check_edge_classes}
\alias{check_edge_classes}
\title{Check edge classes}
\usage{
check_edge_classes(output)
}
\arguments{
\item{output, }{A `list` object returned from the `get_data()` function.}
}
\description{
Check matching edge classes against codebook
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
