% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_testing.R
\name{check_edges_table}
\alias{check_edges_table}
\title{Check `edges_table`}
\usage{
check_edges_table(output)
}
\arguments{
\item{output, }{A `list` object returned from the `get_data()` function.}
}
\description{
Checks `edges_table` element for column and content types.
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
