% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_testing.R
\name{check_metadata}
\alias{check_metadata}
\title{Check \code{metadata} elements}
\usage{
check_metadata(output)
}
\arguments{
\item{output, }{A \code{list} object returned from the \code{get_data()} function.}
}
\description{
Checks metadata for each edge class
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
\keyword{internal}
