% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_testing.R
\name{check_reference}
\alias{check_reference}
\title{Check \code{reference} element}
\usage{
check_reference(output)
}
\arguments{
\item{output, }{A \code{list} object returned from the \code{get_data()} function.}
}
\description{
Check \code{reference} element
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
\keyword{internal}
