% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{core_as_network}
\alias{core_as_network}
\title{Build a \code{network} object from a \code{proto_net}}
\usage{
core_as_network(proto_net)
}
\arguments{
\item{proto_net}{Any \code{{COREnets}} data set obtained via \code{get_data()}.}
}
\value{
\verb{<network>}
}
\description{
Build a \code{network} object from a \code{proto_net}
}
\examples{
drugnet_proto_net <- get_data("drugnet")

core_as_network(drugnet_proto_net)

# currently supported networks ==============================================
core_as_network(get_data("anabaptists"))

core_as_network(get_data("cocaine_smuggling_acero"))

core_as_network(get_data("cocaine_smuggling_jake"))

core_as_network(get_data("cocaine_smuggling_juanes"))

core_as_network(get_data("cocaine_smuggling_mambo"))

core_as_network(get_data("harry_potter_death_eaters"))

core_as_network(get_data("harry_potter_dumbledores_army"))

core_as_network(get_data("montreal_street_gangs"))

core_as_network(get_data("november17"))

core_as_network(get_data("siren"))

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
