% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.corenets_as_edge_df}
\alias{.corenets_as_edge_df}
\title{\code{igraph::as_data_frame()} wrapper

Use in place of \code{igraph::as_data_frame()}/\code{igraph::get.data.frame()}.}
\usage{
.corenets_as_edge_df(igraph, .names = TRUE)
}
\arguments{
\item{igraph}{\code{<igraph>}}

\item{.names}{\code{<lgl>} Whether the \code{"from"} and \code{"to"} columns contain node
\code{"name"}s or numerical indices.}
}
\description{
\code{igraph::as_data_frame()} wrapper

Use in place of \code{igraph::as_data_frame()}/\code{igraph::get.data.frame()}.
}
\keyword{internal}
