% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.corenets_read_file}
\alias{.corenets_read_file}
\title{Read file contents as-is then strip leading/trailing whitespace}
\usage{
.corenets_read_file(file_path, warn = FALSE)
}
\arguments{
\item{file_path}{Path of target file relative to \verb{/inst} (in dev. mode).}

\item{warn}{\code{logical}, Default: \code{FALSE}. Argument passed to \code{base::readLines()}'s.}

\item{...}{Additional arguments passed to or from other methods.}
}
\description{
Read file contents as-is then strip leading/trailing whitespace
}
\keyword{internal}
