% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.has_loops}
\alias{.has_loops}
\title{Does a \code{proto_net} have loops?

Checks whether any \code{from} == \code{to} for the same edge.}
\usage{
.has_loops(proto_net)
}
\arguments{
\item{proto_net}{A network primitive object returned by \code{get_data()}.}
}
\description{
Does a \code{proto_net} have loops?

Checks whether any \code{from} == \code{to} for the same edge.
}
\keyword{internal}
