% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_bimodal}
\alias{.is_bimodal}
\title{Is a \code{proto_net} bimodal?

Checks \code{is_bimodal} in codebook.}
\usage{
.is_bimodal(proto_net)
}
\arguments{
\item{proto_net}{A network primitive object returned by \code{get_data()}.}
}
\description{
Is a \code{proto_net} bimodal?

Checks \code{is_bimodal} in codebook.
}
\keyword{internal}
