% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_multiplex}
\alias{.is_multiplex}
\title{Is a \code{proto_net} multiplex?

Checks if there are duplicate unique rows or multiple \code{edge_class}es.}
\usage{
.is_multiplex(proto_net)
}
\arguments{
\item{proto_net}{A network primitive object returned by \code{get_data()}.}
}
\description{
Is a \code{proto_net} multiplex?

Checks if there are duplicate unique rows or multiple \code{edge_class}es.
}
\keyword{internal}
