% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.print_edge_class_attr}
\alias{.print_edge_class_attr}
\title{Print metadata for all edge_classes}
\usage{
.print_edge_class_attr(.list, ...)
}
\description{
Print metadata for all edge_classes
}
\keyword{internal}
