% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.print_edge_table_attrs}
\alias{.print_edge_table_attrs}
\title{Print edge table attributes}
\usage{
.print_edge_table_attrs(edge_table, wanted)
}
\description{
Print edge table attributes
}
\keyword{internal}
