% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.print_header}
\alias{.print_header}
\title{Print header line for \code{proto_net} object}
\usage{
.print_header(pn, ...)
}
\arguments{
\item{pn}{The object to print, should be "proto_net" class.}

\item{...}{Extra arguments to pass to \code{print()}.}
}
\description{
Print header line for \code{proto_net} object
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
\keyword{internal}
