% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{generate_graph_metadata}
\alias{generate_graph_metadata}
\title{Generate COREnets Graph Metadata}
\usage{
generate_graph_metadata(g, codebook)
}
\arguments{
\item{g, }{An \code{igraph} object.}

\item{codebook, }{A \code{data.frame} with metedata fields for each \code{edge_class}
in the graph object.}
}
\description{
Generate COREnets Graph Metadata
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
\keyword{internal}
