% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{get_codebook_fields}
\alias{get_codebook_fields}
\title{Return Metadata Fields from Codebook}
\usage{
get_codebook_fields(codebook, wanted_edge_class)
}
\arguments{
\item{codebook, }{A \code{data.frame} with metedata fields for each \code{edge_class}}

\item{wanted_edge_class, }{A \code{character} matching a value in the \code{edge_class} variable in the codebook}
}
\description{
Return Metadata Fields from Codebook
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
\keyword{internal}
