% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_calling_functions.R
\name{get_description}
\alias{get_description}
\title{Read Data Description}
\usage{
get_description(dataset)
}
\arguments{
\item{dataset, }{the name of the dataset desired.}
}
\description{
Read Data Description
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
