% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_and_structuring_tools.R
\name{get_node_classes}
\alias{get_node_classes}
\title{Testing for Isolates}
\usage{
get_node_classes(igraph_graph, from_class = "from_class",
  to_class = "to_class")
}
\description{
Testing for Isolates
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
