% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{is_proto_net}
\alias{is_proto_net}
\title{Is this object a proto_net?}
\usage{
is_proto_net(pn)
}
\arguments{
\item{pn}{An R object.}
}
\description{
Logical test for proto_net objects.
}
\author{
Chris Callaghan \email{cjcallag@nps.edu}
}
