% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_calling_functions.R
\name{list_data_sources}
\alias{list_data_sources}
\title{List Available Data}
\usage{
list_data_sources()
}
\description{
List Available Data
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
