% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.proto_net}
\alias{print.proto_net}
\title{proto_net print method}
\usage{
\method{print}{proto_net}(x, ...)
}
\arguments{
\item{x}{The object to print, should be "proto_net" class.}

\item{...}{Extra arguments to pass to \code{print()}.}
}
\description{
proto_net print method
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
