% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_and_structuring_tools.R
\name{read_matrix}
\alias{read_matrix}
\title{Read Matrix}
\usage{
read_matrix(file, sep = ",", top_left_corner = "^\\\\s*?$")
}
\arguments{
\item{file, }{the name of the file which the data are to be read from.}

\item{sep, }{the field separator character.}

\item{top_left_corner}{a regex to test whether or not to automatically assign row and column names.}
}
\description{
Read Matrix
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
