% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{test_dynamic}
\alias{test_dynamic}
\title{Testing for Dynamic Edges}
\usage{
test_dynamic(g)
}
\arguments{
\item{g, }{An \code{igraph} object.}
}
\description{
Testing for Dynamic Edges
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
\keyword{internal}
