% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_and_structuring_tools.R
\name{test_isolates}
\alias{test_isolates}
\title{Testing for Isolates}
\usage{
test_isolates(igraph_graph)
}
\description{
Testing for Isolates
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
