% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{test_loops}
\alias{test_loops}
\title{Testing for Loops}
\usage{
test_loops(g)
}
\arguments{
\item{g, }{An \code{igraph} object.}
}
\description{
Testing for Loops
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
\keyword{internal}
