% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_testing.R
\name{test_output}
\alias{test_output}
\title{Testing for Expected Structures}
\usage{
test_output(output)
}
\arguments{
\item{output, }{A `list` object returned from the `get_data()` function.}
}
\description{
Testing for Expected Structures
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
