% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_and_structuring_tools.R
\name{to_graph}
\alias{to_graph}
\title{Graph Matrices as `igraph` graphs}
\usage{
to_graph(.mat, directed = FALSE, ...)
}
\arguments{
\item{.mat, }{A `matrix` produced by the `to_adj_matrix` function}
}
\description{
Graph Matrices as `igraph` graphs
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
