% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_and_structuring_tools.R
\name{to_one_mode}
\alias{to_one_mode}
\title{Transform to one-mode}
\usage{
to_one_mode(.g, ...)
}
\arguments{
\item{.g, }{An `igraph` object produced by the `to_graph` function}
}
\description{
Transform to one-mode
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
