% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{unnest_edge_class}
\alias{unnest_edge_class}
\title{Unnest Igraph Object by Edge Type}
\usage{
unnest_edge_class(g, edge_class_name)
}
\arguments{
\item{g, }{An igraph object.}

\item{edge_class_name, }{String denoting edge type variable.}
}
\description{
\code{unnest_edge_classs} extracts subnetworks based on the edge_class.
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
\keyword{internal}
