% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_and_structuring_tools.R
\name{unnest_edge_types}
\alias{unnest_edge_types}
\title{Unnest Igraph Object by Edge Type}
\usage{
df <- data.frame(from = c("a", "a", "b"), to = c("a", "b", "c"), from_class = "agent", to_class = "agent", edge_type = c("self", "foo", "bar"))
unnest_edge_types(g, "edge_type")
}
\arguments{
\item{igraph_graph, }{An igraph object.}

\item{edge_type_attr_name, }{String denoting edge type variable.}
}
\description{
`unnest_edge_types` extracts subnetworks based on the edge_type.
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
