% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_testing.R
\name{check_nodes_table}
\alias{check_nodes_table}
\title{Check \code{nodes_table}}
\usage{
check_nodes_table(output)
}
\arguments{
\item{output, }{A \code{list} object returned from the \code{get_data()} function.}
}
\description{
Checks \code{nodes_table} element for column and content types.
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
\keyword{internal}
