% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.corenets_sys_file}
\alias{.corenets_sys_file}
\title{\code{system.file()} wrapper

Use to access files in \code{inst/}!}
\usage{
.corenets_sys_file(file_path)
}
\arguments{
\item{file_path}{Path of target file relative to \code{/inst} (in development mode).}
}
\description{
\code{system.file()} wrapper

Use to access files in \code{inst/}!
}
\keyword{internal}
