% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_and_structuring_tools.R
\name{extract_xlsx}
\alias{extract_xlsx}
\title{Read Relational Excels}
\usage{
listed_df <- get_xlsx("datasets/Afghan Tribal Networks.xlsx")
}
\arguments{
\item{.path, }{A path to the workbook with multiple relational and non-relational network data.}
}
\description{
Read Relational Excels
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
