% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_and_structuring_tools.R
\name{generate_graph_metadata}
\alias{generate_graph_metadata}
\title{Generate COREnets Graph Metadata}
\usage{
df <- data.frame(from = c("a", "a", "b"), to = c("a", "b", "c"), from_class = "agent", to_class = "agent", edge_type = c("self", "foo", "bar"))
g <- graph_from_data_frame(df)
unnest_edge_types(g, "edge_type") %>%
   purrr::map(~.x %>%
              generate_graph_metadata)
}
\arguments{
\item{igraph_graph, }{An igraph object.}
}
\description{
Generate COREnets Graph Metadata
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
