% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{get_codebook_fields}
\alias{get_codebook_fields}
\title{Return Metadata Fields from Codebook}
\usage{
get_codebook_fields(codebook, wanted_edge_type)
}
\description{
Return Metadata Fields from Codebook
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
