% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{get_codebook_fields}
\alias{get_codebook_fields}
\title{Return Metadata Fields from Codebook}
\usage{
get_codebook_fields(codebook, wanted_edge_type)
}
\arguments{
\item{codebook, }{A `data.frame` with metedata fields for each `edge_type`}

\item{wanted_edge_type, }{A `character` matching a value in the `edge_type` variable in the codebook}
}
\description{
Return Metadata Fields from Codebook
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
