% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_calling_functions.R
\name{get_data}
\alias{get_data}
\title{Get Data}
\usage{
get_data(dataset)
}
\arguments{
\item{dataset, }{the name of the dataset desired.}
}
\description{
`get_data` returns a `list` with metadata and network 
object required to generate a sociogram.
}
\author{
Brendan Knapp, \email{brendan.knapp@nps.edu}
}
