% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_calling_functions.R
\name{get_description}
\alias{get_description}
\title{Read Data Description}
\usage{
get_description(dataset)
}
\arguments{
\item{dataset}{\code{<chr>} Name of desired data set.}
}
\description{
Read Data Description
}
\examples{
get_description("drugnet")


}
\seealso{
\code{\link[=list_data_sources]{list_data_sources()}}, \code{\link[=get_data]{get_data()}}
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
