% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{get_node_classes}
\alias{get_node_classes}
\title{Testing for Isolates}
\usage{
get_node_classes(g, from_class = "from_class", to_class = "to_class")
}
\description{
Testing for Isolates
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
