% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{get_node_classes}
\alias{get_node_classes}
\title{Testing for Isolates}
\usage{
get_node_classes(g, from_class = "from_class", to_class = "to_class")
}
\arguments{
\item{g, }{An \code{igraph} object.}

\item{from_class, }{A string pointer to variable containing the node class for the \code{from} column}

\item{to_class, }{A string pointer to variable containing the node class for the \code{to} column}
}
\description{
Testing for Isolates
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
\keyword{internal}
