% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_calling_functions.R
\name{list_data_sources}
\alias{list_data_sources}
\title{List Available Data}
\usage{
list_data_sources()
}
\description{
List Available Data
}
\examples{
list_data_sources()

}
\seealso{
\code{\link[=list_data_sources]{list_data_sources()}}, \code{\link[=get_description]{get_description()}}
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
