% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_and_structuring_tools.R
\name{test_dynamic}
\alias{test_dynamic}
\title{Testing for Dynamic Edges}
\usage{
test_dynamic(igraph_graph)
}
\description{
Testing for Dynamic Edges
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
