% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{test_isolates}
\alias{test_isolates}
\title{Testing for Isolates}
\usage{
test_isolates(g)
}
\arguments{
\item{g, }{An \code{igraph} object.}
}
\description{
Testing for Isolates
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
\keyword{internal}
