% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_and_structuring_tools.R
\name{test_loops}
\alias{test_loops}
\title{Testing for Loops}
\usage{
test_loops(igraph_graph)
}
\description{
Testing for Loops
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
