% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_and_structuring_tools.R
\name{to_matrix}
\alias{to_matrix}
\title{From Dataframe to Adjacency or Incidence Matrix}
\usage{
to_matrix(df, rownames = 1)
}
\arguments{
\item{df, }{A \code{df} produced by the \code{read_xlsx} function.}

\item{rownames, }{A column value for the row that will be reassigned as
rownames and then removed from the data frame.}
}
\description{
\code{to_adj_matrix} returns a \code{matrix} with the values tibble
imported from an excel file.
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
\keyword{internal}
