% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_and_structuring_tools.R
\name{to_one_mode}
\alias{to_one_mode}
\title{Transform to one-mode}
\usage{
to_one_mode(g, project = "rows")
}
\arguments{
\item{g, }{An \code{igraph} object produced by the \code{to_graph} function.}

\item{project, }{Projection type for igraph to extract. "rows" will extract a
one-mode matrix of rows-to-rows, while "columns" will extract a one-mode
matrix of colums-to-columns.}
}
\description{
\code{to_one_mode} takes in an \code{igraph} object, tests if bipartite,
and proceeds to project the network. This evaluation can be also conducted on
one-mode networks, yet no action will occur.`
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
\keyword{internal}
