% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{unnest_edge_types}
\alias{unnest_edge_types}
\title{Unnest Igraph Object by Edge Type}
\usage{
unnest_edge_types(g, edge_type_name)
}
\arguments{
\item{g, }{An igraph object.}

\item{edge_type_name, }{String denoting edge type variable.}
}
\description{
`unnest_edge_types` extracts subnetworks based on the edge_type.
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
